+++ tests/end-to-end/external-solution-integration/Gopkg.lock
@@ -265,7 +265,7 @@
 
 [[projects]]
   branch = "master"
-  digest = "1:f66a2208045ed0985902a3aadfd199de82bf43952f8e906207700c3a6c224cab"
+  digest = "1:5349881b98d7b982111fdb924f807b3e1cfa858310b0df5d9432877ac915f78a"
   name = "github.com/kyma-incubator/compass"
   packages = [
     "components/connector/pkg/graphql/clientset",
@@ -280,7 +280,7 @@
     "tests/end-to-end/pkg/idtokenprovider",
   ]
   pruneopts = "UT"
-  revision = "f8c14f9e8dfbca85a9cd83510da3b88858a4205b"
+  revision = "d70e18706abe6b4aeda91369ba897d61c4e8b91d"
 
 [[projects]]
   digest = "1:833b0dcce87aa06e7f42fbfde09e4bd8dd741f994c722250fc649c3ab68c4cd1"
